[Silver III] Title: 이건 꼭 풀어야 해!, Time: 1148 ms, Memory: 136128 KB -BaekjoonHub

+++ 백준/Silver/17390. 이건 꼭 풀어야 해！/이건 꼭 풀어야 해！.java
@@ -15,11 +15,11 @@ public class Main {
         Q=Integer.parseInt(input[1]);
 
 
-        int [] sequence=new int[N];
+        short [] sequence=new short[N];
         int [] sum=new int[N+1];
         input=br.readLine().split(" ");
         for (int i = 0; i < N; i++) {
-            sequence[i]=Integer.parseInt(input[i]);
+            sequence[i]=Short.parseShort(input[i]);
         }
         Arrays.sort(sequence);
         sum[0]=0;
